requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.51.2
google-generativeai==0.8.3
